# dymer-webserver/swagger_webserver.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "DYMER API",
    "description": "Documentation of <b>Webserver</b> module."
  },
  "host": "localhost:0000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read_pets": "read your pets",
        "write_pets": "modify pets in your account"
      }
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/forms/api/v1/form/mongostate": {
      "get": {
        "tags": [
          "Forms"
        ],
        "description": "",
        "parameters": [
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          },
          {
            "name": "dmts",
            "in": "query",
            "type": "string"
          },
          {
            "name": "query",
            "in": "query",
            "type": "string"
          },
          {
            "name": "act",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/forms/api/v1/form/dettagliomodel": {
      "get": {
        "tags": [
          "Forms"
        ],
        "description": "",
        "parameters": [
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          },
          {
            "name": "dmts",
            "in": "query",
            "type": "string"
          },
          {
            "name": "query",
            "in": "query",
            "type": "string"
          },
          {
            "name": "act",
            "in": "query",
            "type": "string"
          },
          {
            "name": "req",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/forms/api/v1/form/": {
      "get": {
        "tags": [
          "Forms"
        ],
        "description": "",
        "parameters": [
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          },
          {
            "name": "dmts",
            "in": "query",
            "type": "string"
          },
          {
            "name": "query",
            "in": "query",
            "type": "string"
          },
          {
            "name": "act",
            "in": "query",
            "type": "string"
          },
          {
            "name": "req",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Forms"
        ],
        "description": "",
        "parameters": [
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          },
          {
            "name": "dmts",
            "in": "query",
            "type": "string"
          },
          {
            "name": "query",
            "in": "query",
            "type": "string"
          },
          {
            "name": "act",
            "in": "query",
            "type": "string"
          },
          {
            "name": "req",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/forms/api/v1/form/content/{entype}/{fileid}": {
      "get": {
        "tags": [
          "Forms"
        ],
        "description": "",
        "parameters": [
          {
            "name": "entype",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "fileid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          },
          {
            "name": "dmts",
            "in": "query",
            "type": "string"
          },
          {
            "name": "query",
            "in": "query",
            "type": "string"
          },
          {
            "name": "act",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/forms/api/v1/form/create": {
      "post": {
        "tags": [
          "Forms"
        ],
        "description": "",
        "parameters": [
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          },
          {
            "name": "dmts",
            "in": "query",
            "type": "string"
          },
          {
            "name": "query",
            "in": "query",
            "type": "string"
          },
          {
            "name": "act",
            "in": "query",
            "type": "string"
          },
          {
            "name": "req",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/forms/api/v1/form/addAsset": {
      "post": {
        "tags": [
          "Forms"
        ],
        "description": "",
        "parameters": [
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          },
          {
            "name": "dmts",
            "in": "query",
            "type": "string"
          },
          {
            "name": "query",
            "in": "query",
            "type": "string"
          },
          {
            "name": "act",
            "in": "query",
            "type": "string"
          },
          {
            "name": "req",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/forms/api/v1/form/update": {
      "post": {
        "tags": [
          "Forms"
        ],
        "description": "",
        "parameters": [
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          },
          {
            "name": "dmts",
            "in": "query",
            "type": "string"
          },
          {
            "name": "query",
            "in": "query",
            "type": "string"
          },
          {
            "name": "act",
            "in": "query",
            "type": "string"
          },
          {
            "name": "req",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/forms/api/v1/form/updatestructure": {
      "post": {
        "tags": [
          "Forms"
        ],
        "description": "",
        "parameters": [
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          },
          {
            "name": "dmts",
            "in": "query",
            "type": "string"
          },
          {
            "name": "query",
            "in": "query",
            "type": "string"
          },
          {
            "name": "act",
            "in": "query",
            "type": "string"
          },
          {
            "name": "req",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/forms/api/v1/form/updateAsset": {
      "post": {
        "tags": [
          "Forms"
        ],
        "description": "",
        "parameters": [
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          },
          {
            "name": "dmts",
            "in": "query",
            "type": "string"
          },
          {
            "name": "query",
            "in": "query",
            "type": "string"
          },
          {
            "name": "act",
            "in": "query",
            "type": "string"
          },
          {
            "name": "req",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/forms/api/v1/form/{id}": {
      "delete": {
        "tags": [
          "Forms"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          },
          {
            "name": "dmts",
            "in": "query",
            "type": "string"
          },
          {
            "name": "query",
            "in": "query",
            "type": "string"
          },
          {
            "name": "act",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/forms/api/v1/form/{id}/{fid}": {
      "delete": {
        "tags": [
          "Forms"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "fid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          },
          {
            "name": "dmts",
            "in": "query",
            "type": "string"
          },
          {
            "name": "query",
            "in": "query",
            "type": "string"
          },
          {
            "name": "act",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}